+++ packages/terra-date-picker/src/DatePicker.jsx
@@ -116,6 +116,11 @@ const propTypes = {
    * The value must be in the `YYYY-MM-DD` format or the all-numeric date format based on the locale.
    */
   value: PropTypes.string,
+  /**
+   * @private
+   * Prop to show inline version of date picker component.
+   */
+  isInline: PropTypes.bool,
 };
 
 const defaultProps = {
@@ -137,6 +142,7 @@ const defaultProps = {
   required: false,
   disableButtonFocusOnClose: false,
   selectedDate: undefined,
+  isInline: false,
 };
 
 class DatePicker extends React.Component {
@@ -146,7 +152,7 @@ class DatePicker extends React.Component {
     const activeBreakpointOnMount = activeBreakpointForSize(window.innerWidth);
     this.state = {
       selectedDate: DateUtil.defaultValue(props),
-      showPortalPicker: activeBreakpointOnMount === 'tiny' || activeBreakpointOnMount === 'small',
+      showPortalPicker: !this.props.isInline && (activeBreakpointOnMount === 'tiny' || activeBreakpointOnMount === 'small'),
       prevPropsSelectedDate: props.value || props.selectedDate,
     };
 
@@ -219,7 +225,7 @@ class DatePicker extends React.Component {
   }
 
   handleBreakpointChange(activeBreakpoint) {
-    const showPortalPicker = activeBreakpoint === 'tiny' || activeBreakpoint === 'small';
+    const showPortalPicker = !this.props.isInline && (activeBreakpoint === 'tiny' || activeBreakpoint === 'small');
 
     if (this.state.showPortalPicker !== showPortalPicker) {
       this.setState({ showPortalPicker });
@@ -393,6 +399,7 @@ class DatePicker extends React.Component {
       required,
       selectedDate,
       value,
+      isInline,
       ariaLabel,
       ...customProps
     } = this.props;
@@ -440,6 +447,7 @@ class DatePicker extends React.Component {
           <ReactDatePicker
             {...customProps}
             withPortal={this.state.showPortalPicker}
+            inline={isInline}
             selected={selectedDateInPicker}
             value={formattedValue}
             onBlur={this.handleBlur}