+++ packages/terra-date-picker/src/terra-dev-site/doc/example/CalendarFilter/CalendarFilterDefault.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import CalendarFilter from 'terra-date-picker/lib/CalendarFilter';
+
+const CalendarFilterDefault = () => (
+  <div>
+    <p>Opens to current date if no selected date is provided.</p>
+    <CalendarFilter
+      id="default"
+    />
+  </div>
+);
+
+export default CalendarFilterDefault;